Add lightweight analytics pixel to dapr website
This analytics pixel provides high-level web traffic insights for project maintainers so they can better understand how the dapr website is being used. This pixel does not use cookies (or JavaScript at all) and is GDPR-compliant. It is a clear 1x1 pixel that won't affect page rendering or load times.

Signed-off-by: Arjun Devarajan <arjun@scarf.sh>